Changed to integer output and added test script

+++ include/createSampleImage.h
@@ -1,4 +1,4 @@
 #include "platformDefines.h"
 
-EXPORT void createImageEMCCD(double *binnedImage, const double potentialAtomLocations[][2], unsigned short cameraCoords, double *truth, int atomCount, int approximationSteps);
-EXPORT void createImageCMOS(double *binnedImage, const double potentialAtomLocations[][2], unsigned short cameraCoords, double *truth, int atomCount, int approximationSteps);
+EXPORT void createImageEMCCD(int *binnedImage, const double potentialAtomLocations[][2], unsigned short cameraCoords, double *truth, int atomCount, int approximationSteps);
+EXPORT void createImageCMOS(int *binnedImage, const double potentialAtomLocations[][2], unsigned short cameraCoords, double *truth, int atomCount, int approximationSteps);
